style(code): 格式化代码，清理无用代码

+++ Round.SDK/Helper/ZipHelper.cs
@@ -7,10 +7,7 @@ public class ZipHelper
     public static void CreateZipFile(string sourceFolder, string zipPath)
     {
         // 如果ZIP文件已存在，先删除
-        if (File.Exists(zipPath))
-        {
-            File.Delete(zipPath);
-        }
+        if (File.Exists(zipPath)) File.Delete(zipPath);
 
         // 创建ZIP文件
         ZipFile.CreateFromDirectory(sourceFolder, zipPath);
@@ -19,10 +16,7 @@ public class ZipHelper
     public static void ExtractZipFile(string file, string extractDir)
     {
         // 如果目录已存在，退出
-        if (Directory.Exists(extractDir))
-        {
-            return;
-        }
+        if (Directory.Exists(extractDir)) return;
 
         Directory.CreateDirectory(extractDir);
